added Fireball from the original Limited base set.  Everything seems to work for the Human.  The AI doesn't do anything terribly clever - just finishes human off if it can.  But this is a good start.
There are also some kinks with the flow of selecting a player or creature twice.  But others can add their input.

Also, there is a lot of debug code there that I left in on purpose.  Once we are confident the kinks are out, or before an upcoming release, it can be commented/removed.  Enjoy!